made a sperate helper file from EnglishKeyboardIme and 20 percent increase in key heigth for tablet mode
