+++ app/src/main/res/values-w1240dp/dimens.xml
@@ -6,8 +6,8 @@
     <dimen name="toolbar_height">48dp</dimen>
     <dimen name="toolbar_icon_height">44dp</dimen>
     <dimen name="clip_drag_icon_height">48dp</dimen>
-    <dimen name="key_height">72dp</dimen>
-    <dimen name="key_height_landscape">48dp</dimen>
+    <dimen name="key_height">80dp</dimen>
+    <dimen name="key_height_landscape">56dp</dimen>
     <dimen name="vertical_correction">-12dp</dimen>
     <dimen name="clip_pin_size">32dp</dimen>
